+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Added config for logging actions and reporting
 - Suppress packet handling errors from blacklisted IPs
 - Allow later Fabric API versions
+- Enforce dedicated server only